// clients/socialService/models/SocialPostAttachmentCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SocialPostAttachmentCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    notes?: string | null;
    title?: string | null;
    author?: string | null;
    isFolder?: boolean;
    fileName?: string | null;
    abstract?: string | null;
    keyWords?: string | null;
    validResponse?: boolean;
    parentFileUploadId?: string | null;
    filePath?: string | null;
    file?: Blob | null;
    socialPostId?: string | null;
};