// clients/socialService/models/SocialPostCommentCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SocialPostCommentCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    message: string;
    parentCommentId?: string | null;
    socialProfileId?: string | null;
    socialFeedPostId?: string | null;
    socialPostId?: string | null;
};